Create/Edit Methods
I created a new a method called 'is_space_free' that returns true if a space on the board is free and false if it's full.  This was originally part of the 'is_user_input_valid' method but I decided to break that method up into 2 seperate methods.

I also updated the 'move_confirmation' method to be more specific and prints a message if the user's choice is already taken.