Playing_Game Method Added
I added a new method called 'playing_game' that marks the board and then changes the marker from 'X' to 'O' and vice versa. I also updated the 'change_marker' method to include a new class variable called '@pos_arrow', which can easily be changed from 1 to -1 by multiplying the by -1. So if the '@pos_arrow' is equal to 1, then the 'get_marker' method would return 'X' and if '@pos_arrow' is equal to -1, then the 'get_marker' method would return 'O'.