Player Wins Method
Created a new method that returns true/false if a player has won or not. I also added a new class variable called '@WINNING_COMBINATIONS' that is a nested array that holds each winning combination possible on the board. This nested array is iterated through in the 'player_wins' method to check to see if the player has a match.