New Winning_Player Method
I added a new method called 'winning_player' that returns the marker of which player won the game. I used the same loop in the 'player_wins' method but instead of returning a boolean, it returns either the winning player's marker or the String 'No winner'. I updated the 'player_wins' method to run return true/false whether or not the 'winning_player' method returns 'No winner'.