Integrated Minimax AI to Game
I added the minimax function to the game, but after doing a few test games, I found that the computer is still beatable. If the user makes their first move in the center of the board, the computer will always make a move directly above the center spot. Then if the user makes their next move to the left or right of the center spot, they can win the game.