Updated AI
I updated the 'cpu-pick' method to always pick the center spot as the CPU's first move if it's open. This makes it more difficult for the user to win. I also updated the 'mark_board' method to verify if the space being marked was free because I noticed a bug in the game that would allow the CPU to choose the '0' spot if the human player chooses the '0' spot as their first move.