Updated Minimax Function
I updated the minimax function to not take in the player as a variable. Instead it takes in 2 variables called 'depth' and 'isMaximizing'. Depth is an integer that represents the level of recursion and 'isMaximizing' is a boolean. I also seperated the prior minimax function into two functions, the new minimax and another function called 'best_move_to_make'. So the new minimax function returns the best score possible out of the available spots on the board and the best_move_to_make function returns where the CPU will move on the board.